new chnages

+++ src/app/home/home.component.ts
@@ -1,12 +1,12 @@
 import { Component, OnInit, ViewChild } from "@angular/core";
 import { MatTableDataSource } from "@angular/material/table";
 import { SelectionModel } from "@angular/cdk/collections";
-import { contactData, ELEMENT_DATA } from "./home.model";
+import { contactData } from "./home.model";
 import { ContactService } from "../shared/contact.service";
-import { isNgTemplate } from "@angular/compiler";
 import { MatSort, MatPaginator } from "@angular/material";
 import { MatDialog, MatDialogConfig } from "@angular/material";
 import { ContactFormComponent } from "../contact-form/contact-form.component";
+import { NotificationService } from "../shared/notification.service";
 
 @Component({
   selector: "app-home",
@@ -19,7 +19,8 @@ export class HomeComponent implements OnInit {
   searchKey: string;
   constructor(
     private contactService: ContactService,
-    private dialog: MatDialog
+    private dialog: MatDialog,
+    private notificationService: NotificationService
   ) {}
   displayedColumns: string[] = [
     "First Name",
@@ -40,11 +41,9 @@ export class HomeComponent implements OnInit {
           ...item.payload.val()
         };
       });
-      console.log("array from home array", new MatTableDataSource(array));
       this.dataSource = new MatTableDataSource<contactData>(array);
-      console.log("array from home array after", this.dataSource);
-      this.dataSource.sort = this.sort;
-      this.dataSource.paginator = this.paginator;
+      this.dataSource.sort = this.sort; // sorting feature
+      this.dataSource.paginator = this.paginator; // pagination feature
     });
   }
 
@@ -76,6 +75,9 @@ export class HomeComponent implements OnInit {
   }
 
   onDelete($key) {
-    this.contactService.deleteContact($key);
+    if (confirm("Are you sure to delete this contact")) {
+      this.contactService.deleteContact($key);
+      this.notificationService.success("Contact deleted successfully");
+    }
   }
 }
